Create Capsule-people-age
A new med is out in the market which helps in treating auto-immune disease in humans these are very common in zero to 10 yrs kid that or greater than or equal to 81 of senior citizon. These age group are considered as high risk people. In a single day we can provide both high and non high risk. Each person requires only one capsule in a day. There are L capsules distributed in a day .The high risk people are supposed to be consider first. Find the number of days requested to medicate all the people.

INPUT
no.of people
no.of capsule
ages
SAMPLE INPUT
5
2
81
22
34
55
65
OUTPUT
no.of days
SAMPLE OUTPUT
3