Create play with numbers
Mikes likes to play with numbers. His friends are also good with numbers and often plays mathematical games. They made a small game where they will spell the last digit of a factorial of a number other than 0. Let say the given number is 5, so 5! (5 factorial) will be
5 4 3* 2* 1= 120.
Here 0 is the last digit. But, we don't want 0, we want a
number other than 0.
Then the last digit is 2. This is what we have to output
Input Format

First input-Accept the value of integer
Constraints

•1<=input number<105
Output Format

The output is the last nonzero digit, of a factorial of the input number
Sample Input 0

9
Sample Output 0

8
Sample Input 1

5
Sample Output 1

2