Create Digital root of ancient scroll number
In a small village, the townsfolk discovered a number N written on an ancient scroll. They believe that to unlock its secrets, they must reduce the number N to a single digit by summing its digits repeatedly. Your task is to help the townsfolk unlock the scroll by returning an integer value representing the reduced single digit of the number N.

Input Format

An integer value representing the number N written on the ancient scroll.

Constraints

NA

Output Format

Return an integer value representing the reduced single digit of the number N.

Sample Input 0

38
Sample Output 0

2
Explanation 0

The number on the ancient scroll is 38. To reduce this number, we will add the digits:

3 + 8 = 11

Since the number 11 is not a single digit, we add its digits again:

1 + 1 = 2

Since 2 is a single-digit number, it is the final result.

Sample Input 1

41
Sample Output 1

5
Explanation 1

The number on the ancient scroll is 41, to reduce this number we will add the digits 4 + 1, which is a single digit number. Therefore 5 is returned as the output.