Create Sum of Highest Square and Highest Cube
Mike is given an integer number N. He is asked to find out all those numbers within the range of 1 to N (inclusive), which are:

Either Square of any number or;
Cube of any number
Find and return the sum of the highest square and the highest cube in the range of 1 to N (inclusive).

Input Format

An integer value N.

Constraints

NA

Output Format

Return an integer value representing the sum of the highest square and the highest cube.

Sample Input 0

10
Sample Output 0

17
Explanation 0

In this scenario, we have a range of 1 to 10. In this range, there are below numbers which follow the logic:

1: square and cube of 1
4: square of 2
8: cube of 2
9: square of 3
The highest square is 9

The highest cube is 8

Sum of these numbers = 8 + 9 = 17. Therefore, 17 is returned as the output.

Sample Input 1

64
Sample Output 1

128
Explanation 1

In this scenario, we have a range of 1 to 64. In this range, there are below numbers which follow the logic:

1: square and cube of 1
4: square of 2
8: cube of 2
9: square of 3
16: square of 4
25: square of 5
27: cube of 3
36: square of 6
49: square of 7
64: square of 8 and cube of 4
The highest square is 64

The highest cube is 64

Sum of these numbers = 64 + 64 = 128. Therefore, 128 is returned as the output.