Create Count digits in string
To create a profile on a social media account "ThePastBook", the user needs to enter a string value in the form of a username. The username should consist of only characters tagged a-z. If the user enters an incorrect string containing digits the system automatically identifies the number of digits in the string and removes them.

Write an algorithm to help the system identify the count of digits present in the username.

Example

Input

rah23ul

Output

2

Explanation

There are two digits in the username, '2' and '3'. The total count of digits is two (2). Hence 2 is the output.

Input Format

The first line of input consists of a string - userName, representing the username.

Constraints

1 <= Length of S<= 10^5

Output Format

The output prints an integer representing the total count of digits present in the username. If no digits are present output 0.

Sample Input 0

hello 23 world
Sample Output 0

2
Sample Input 1

genera32dfs2342
Sample Output 1

6