Create Credit Score
The new bank, "YoursPay", has a list of N customers' bank account balances. The list consists of both positive and negative balances. The positive balance signifies the current year's customers, and the negative balance signifies last year's customers. The bank has decided to offer shortlisted customers credit scores to their credit cards. The credit score will be the sum of the two balances from the list with the smallest product when multiplied. If the credit score is positive then the credit will be provided to the current year's customer, otherwise, it will go to the last year's customer. Write an algorithm to find the credit score.

Input Format

The first line of input consists of an integer - numCustomers, representing the number of banking customers (N). The second line of input consists of N space-separated integers - balance0, balance1, ..... balanceN-1 representing the customers' bank balances.

Constraints

2 <= numCustomers <= 1000 -10^9 <= balancei <= 10^9 0 <= i < numCustomers

Output Format

Print an integer representing the credit score.