Create DP - fibonacci sum
The games development company "FunGames" has developed a balloon shooter game. The balloons are arranged in a linear sequence and each balloon has a number associated with it. The numbers on the balloons are in the fibonacci series. In the game, the player shoots 'k' balloons. The player's score is the sum of numbers on the 'k' balloons. Write an algorithm to generate the player's score. Example Input 7 Output 20 Explanation The fibonacci sum is 0+1+1+2+3+5+8=20

Input Format

The input consists of an integer - numBalloons representing the total number of balloons shot by the player (k).

Constraints

0 < numBalloons <= 10^6

Output Format

Print an integer value representing the player's score. If no balloons are shot then print 0.

Sample Input 0

7
Sample Output 0

20
Sample Input 1

6
Sample Output 1

12