Create Card pyramid
Arrangement of cards used for building pyramids are shown in the following image:

Fig 2: level 2 pyramid

You are required to implement the following function :

int cardspyramid(int n);

The function accepts an integer ‘n’ as its argument. The integer ‘n’ denotes the level of the pyramid. You are required to calculate the numbers of cards, required to build a pyramid of level ‘n’, and return the number of cards

Note

Return -1 , if ‘n’ = 0

Example

Input

2

Output

7

Explanation

Cards required to build a pyramid of level 1 are 2, adding 1 more level to the pyramid will require 5 more cards, thus a total of cards are needed to build a pyramid of level 2, hence 7 is returned.

Input Format

An integer representing the number of levels in the pyramid.

Constraints

The number of cards required to build a pyramidal of level 1 is 2.

Output Format

An integer representing the number of cards required to build the pyramid.

Sample Input 0

3
Sample Output 0

15
Sample Input 1

2
Sample Output 1

7