Create String palindrome
A palindrome is a sequence of characters that has the property of reading the same in either direction. You are given a function.

Char* Convertpalindrome ( char* str );

The function accepts a string ‘str’ .Implement the function to find and return the minimum characters required to append at the end of string ‘str’ to make it a palindrome.

Assumption

The string will contain only lowercase English alphabets.

The length of the string is greater than equal to 1.

Note

If the string is already palindrome then return “NULL”.

You have to find the minimum characters required to append at the end of the string to make it a palindrome.

Example1

Input

abcdc

Output

ba

Explanation

If we append 'ba' at the end of the string 'abcdc', it becomes 'abcdcba' (i.e. A palindrome string).

Example2

Input

abcde

Output

dcba

Explanation

If we append 'dcba' at the end of the string 'abcde', it becomes 'abcdedcba' (i.e. A palindrome string).

Input Format

Input is a single string.

Constraints

1 < S < 1000

Output Format

The output contains the characters to make the input string a palindrome.

Sample Input 0

abcdc
Sample Output 0

ba
Sample Input 1

fehgh
Sample Output 1

ef