Create Minimum sum - Greedy
You are given two integer arrays A and B of length N on which you have to perform below operation.

In one operation, you can swap any two elements of 'A' or any two elements of B

Your task is to find and return an integer value representing the minimum possible sum of A[i]* B[i] after performing the above operation any number of times.

Note

The operation can also be performed number of times.

Input Format

An integer value S representing the total number of daily steps.
An integer array A
An integer array B
Constraints

NA

Output Format

Return an integer value representing the minimum possible sum of A[i]*B[i] after performing the above operation any number of times.

Sample Input 0

4
1 4 3 2
1 4 3 4
Sample Output 0

25
Explanation 0

Here, A = (1,4,3,2) and B = (1,4,3,4). To minimize the sum, we can swap the first two elements of A, ie, 4 and 1. The array will now become (4,1,3,2). The sum obtained will be 25, which is the minimum. Hence, 25 is returned as the output.

Sample Input 1

3
1 2 3
4 5 6
Sample Output 1

28