Create E comm
The delivery request of an e-commerce website is depicted in the form of strings tagged 'a-z' or 'A-Z'. Each character of the string represents the request code. The ascii value of the character represents the number of products that need to be delivered for that respective code. Due to limited delivery windows, the website has decided to deliver the products whose codes represent the largest and the smallest ascii values in the string. The company needs to find out the total number of products delivered.
Write an algorithm to find out the total number of products delivered
Input Format

A string value
Constraints

NA
Output Format

An intger value
Sample Input 0

aAkliP
Sample Output 0

173
Explanation 0

The ascii value of the characters in the string are as follows:
a = 97
A = 65
k = 107
l = 108
i = 105
p = 80
The sum of the smallest and the largest size values is 65 + 108 = 173. Hence the output is 173.
Sample Input 1

Akil
Sample Output 1

173