Create Min sum game
Problem Statement:
Ram is a boy who is very interested in playing new games, so his mother took him to a puzzle game and asked him to solve it. His mother gave some instructions to solve the puzzle game. Now he has to follow all the given instructions while solving the game. His mother explained the step by step rule of the game and told that first you have to take n elements, after that you have to sum the digits of that n taken numbers.
After adding the digits of the taken number, He has to check that the sum of which digit numbers is minimum. Now Ram has to find the minimum sum of digits of the number which he had taken in the beginning of the game.
Write a proper algorithmic code to solve that puzzle game

Constraints :
1<=n<= 10^5

Input :
4 = number of elements
21 2 30 33 elements

Output :
2